Added a generic object pool
Added a generic object pool that can be used by anyone. I made it so that it because I need a pool for the enemies but it is set up as a generic object pool. Use the IReturnSelfObject interface if you want the object in the pool to return itself (for example a bullet returns itself when it hits an object or after 10 seconds). Just implement the interface and it should work.